:whale: Add docker files to build and execute images
* Added `Dockerfile` and `docker-compose.yml`. The docker image is based on python.
* Added `install-default-environment.sh` for installing all required dependencies and configuring the environment in the docker image.
* The docker image create a git repository inside the container for testing purposes. The graph of the git repo can be found in `install-default-environment.sh`.
* Exported python modules name in `requirements.txt` for better cross-compatibility :computer: .
* Added `default-gitconfig.ini` for the docker image.
* :bug: Fix bug that appears when no output file were given.

+++ commity.py
@@ -54,7 +54,7 @@ if args.branch not in repo.branches:
 	exit(-3)
 
 # If an output has been given and the file already exist, remove it:
-if os.path.exists(args.output) and os.path.isfile(args.output):
+if args.output is not None and os.path.exists(args.output) and os.path.isfile(args.output):
 	os.remove(args.output)
 
 log("On branch " + args.branch, end="\n\n")
